refactor(fe): split appearance preview components
