+++ frontend/core/unit/DashboardThread/Appearance/KanbanLayout/ItemCardLayout.tsx
@@ -1,14 +1,14 @@
 import { KANBAN_CARD_LAYOUT } from '~/const/layout'
 import useTrans from '~/hooks/useTrans'
-import CommentSVG from '~/icons/Comment'
-import UpvoteSVG from '~/icons/Upvote'
 import CheckLabel from '~/widgets/CheckLabel'
 
 import { FIELD } from '../../constant'
 import useKanban from '../../logic/useKanban'
 import SavingBar from '../../SavingBar'
 import SectionLabel from '../../SectionLabel'
-import useSalon, { cnMerge } from './salon/item_card_layout'
+import FullCardPreview from './FullCardPreview'
+import useSalon from './salon/item_card_layout'
+import SimpleCardPreview from './SimpleCardPreview'
 
 const KANBAN_CARD_LAYOUT_OPTIONS = [
   {
@@ -20,63 +20,6 @@ const KANBAN_CARD_LAYOUT_OPTIONS = [
     titleKey: 'dsb.appearance.kanban.card.option.full',
   },
 ] as const
-
-function SimplePreview({ isActive }: { isActive: boolean }) {
-  const s = useSalon()
-
-  return (
-    <div className={cnMerge(s.block, isActive && s.blockActive)}>
-      <div className={s.frame}>
-        <div className={s.header}>
-          <div className={cnMerge(s.bar, s.titleBar)} />
-          <div className={cnMerge(s.bar, s.bodyBar)} />
-        </div>
-
-        <div className={s.footer}>
-          <div className={s.footerLeft}>
-            <UpvoteSVG className={s.icon} />
-            <CommentSVG className={s.commentIcon} />
-          </div>
-          <div className={cnMerge(s.bar, s.simpleMetric)} />
-        </div>
-      </div>
-    </div>
-  )
-}
-
-function FullPreview({ isActive }: { isActive: boolean }) {
-  const s = useSalon()
-
-  return (
-    <div className={cnMerge(s.block, isActive && s.blockActive)}>
-      <div className={s.frame}>
-        <div className={s.header}>
-          <div className={cnMerge(s.bar, s.titleBar)} />
-          <div className={s.headerRow}>
-            <div className={cnMerge(s.bar, s.bodyBar)} />
-            <div className={cnMerge(s.bar, s.sideBar)} />
-          </div>
-        </div>
-
-        <div className={s.footer}>
-          <div className={s.footerLeft}>
-            <UpvoteSVG className={s.icon} />
-            <div className={s.avatarList}>
-              <div className={s.userAvatar} />
-              <div className={cnMerge(s.userAvatar, 'opacity-30')} />
-              <div className={cnMerge(s.userAvatar, 'opacity-20')} />
-            </div>
-          </div>
-
-          <div className={s.footerRight}>
-            <CommentSVG className={s.commentIcon} />
-            <div className={cnMerge(s.bar, s.tinyMetric)} />
-          </div>
-        </div>
-      </div>
-    </div>
-  )
-}
 
 export default function ItemCardLayout() {
   const s = useSalon()
@@ -103,9 +46,9 @@ export default function ItemCardLayout() {
               onClick={() => edit(value, FIELD.KANBAN_CARD_LAYOUT)}
             >
               {value === KANBAN_CARD_LAYOUT.SIMPLE ? (
-                <SimplePreview isActive={isActive} />
+                <SimpleCardPreview isActive={isActive} />
               ) : (
-                <FullPreview isActive={isActive} />
+                <FullCardPreview isActive={isActive} />
               )}
               <CheckLabel title={t(titleKey)} active={isActive} top={2} />
             </button>
