+++ frontend/core/unit/DashboardThread/Appearance/PostLayout/spec.d.ts
@@ -1,0 +1,4 @@
+export type TPreviewProps = {
+  isActive: boolean
+  compact?: boolean
+}
