+++ frontend/core/unit/DashboardThread/Appearance/Theme/GlassOpacity.tsx
@@ -1,12 +1,10 @@
-import { useRef, useState } from 'react'
-
 import useTrans from '~/hooks/useTrans'
 import type { TSpace } from '~/spec'
-import RangeInput from '~/widgets/RangeInput'
 
 import { FIELD } from '../../constant'
 import useSettingRowSalon from './salon/setting_row'
 import type { TThemePresetOverwrite } from './spec'
+import ThemeRangeInput from './ThemeRangeInput'
 
 type TProps = {
   selectedOverwrite: TThemePresetOverwrite
@@ -15,51 +13,6 @@ type TProps = {
   onThemePresetSchedule: (patch: Partial<TThemePresetOverwrite>) => void
   onThemePresetFlush: () => void
 } & TSpace
-
-type TGlassOpacityRangeProps = {
-  value: number
-  valueLabel: string
-  getPatch: (value: number) => Partial<TThemePresetOverwrite>
-  onThemePresetPreview: (patch: Partial<TThemePresetOverwrite>) => void
-  onThemePresetSchedule: (patch: Partial<TThemePresetOverwrite>) => void
-  onThemePresetFlush: () => void
-}
-
-const GlassOpacityRange = ({
-  value,
-  valueLabel,
-  getPatch,
-  onThemePresetPreview,
-  onThemePresetSchedule,
-  onThemePresetFlush,
-}: TGlassOpacityRangeProps) => {
-  const initialValueRef = useRef(value)
-  const [displayGaussBlur, setDisplayGaussBlur] = useState(initialValueRef.current)
-
-  return (
-    <RangeInput
-      value={displayGaussBlur}
-      valueLabel={valueLabel}
-      min={50}
-      max={100}
-      step={0.1}
-      unit='%'
-      top={0}
-      aria-label={valueLabel}
-      onChange={(value) => {
-        const patch = getPatch(value)
-
-        setDisplayGaussBlur(value)
-        onThemePresetPreview(patch)
-        onThemePresetSchedule(patch)
-      }}
-      onChangeEnd={(value) => {
-        onThemePresetSchedule(getPatch(value))
-        onThemePresetFlush()
-      }}
-    />
-  )
-}
 
 export default function GlassOpacity({
   selectedOverwrite,
@@ -89,10 +42,12 @@ export default function GlassOpacity({
 
           <div className='grow' />
           <div className={s.rangeGroup}>
-            <GlassOpacityRange
+            <ThemeRangeInput
               key={`${gaussBlurField}-${gaussBlur}`}
               value={gaussBlur}
               valueLabel={t('dsb.appearance.theme.glass_opacity.title')}
+              min={50}
+              max={100}
               getPatch={getGaussBlurPatch}
               onThemePresetPreview={onThemePresetPreview}
               onThemePresetSchedule={onThemePresetSchedule}
