+++ frontend/core/unit/DashboardThread/Appearance/Theme/MiniColorBar.tsx
@@ -1,0 +1,8 @@
+type TProps = {
+  className?: string
+  color: string
+}
+
+export default function MiniColorBar({ color, className }: TProps) {
+  return <span className={className} style={{ backgroundColor: color }} />
+}
